# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClinicalFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, scheduler, ci, student]
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deactivateUser
      tags: [users]
      summary: Deactivate user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deactivated

  # ── STUDENTS ──────────────────────────────────────────────────────────────
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students with filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: needsMakeup
          in: query
          schema:
            type: boolean
        - name: hospitalId
          in: query
          schema:
            type: string
        - name: caseId
          in: query
          schema:
            type: string
        - name: minHours
          in: query
          schema:
            type: number
        - name: maxHours
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentDetail"
  /students/{id}/passport:
    get:
      operationId: getStudentPassport
      tags: [students]
      summary: Get student clinical passport summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clinical passport
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicalPassport"
  /students/{id}/attendance:
    get:
      operationId: getStudentAttendance
      tags: [students]
      summary: Get student attendance history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /students/{id}/hours:
    get:
      operationId: getStudentHours
      tags: [students]
      summary: Get student total duty hours
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hours summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HoursSummary"

  # ── HOSPITALS ─────────────────────────────────────────────────────────────
...
  /slots:
    get:
      operationId: listSlots
      tags: [slots]
      summary: List available duty slots
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, cancelled]
        - name: isMakeup
          in: query
          schema:
            type: boolean
        - name: hospitalId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DutySlot"
    post:
      operationId: createSlot
      tags: [slots]
      summary: Create available duty slot (Scheduler)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DutySlotInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutySlot"
  /slots/{id}:
    patch:
      operationId: updateSlot
      tags: [slots]
      summary: Edit duty slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DutySlotInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutySlot"
    delete:
      operationId: cancelSlot
      tags: [slots]
      summary: Cancel duty slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
  /slots/{id}/apply:
    post:
      operationId: applyForSlot
      tags: [slots]
      summary: Student applies for a duty slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutyApplication"
  /slots/{id}/applications:
    get:
      operationId: listSlotApplications
      tags: [slots]
      summary: List applications for a slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DutyApplication"
  /slots/{id}/applications/{appId}:
    patch:
      operationId: reviewSlotApplication
      tags: [slots]
      summary: Approve or reject a slot application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: appId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationReview"
      responses:
        "200":
          description: Application reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DutyApplication"

  # ── RECOMMENDATIONS ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, role, firstName, lastName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, scheduler, ci, student]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        ciProfile:
          $ref: "#/components/schemas/CIProfile"
    UserInput:
      type: object
      required: [email, password, role, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [admin, scheduler, ci, student]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        studentNumber:
          type: string
        yearLevel:
          type: integer
        section:
          type: string
        program:
          type: string
        academicYear:
          type: string
        employeeId:
          type: string
        specialization:
          type: string
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        isActive:
          type: boolean
        section:
          type: string
        yearLevel:
          type: integer
    StudentProfile:
      type: object
      required: [id, userId, studentNumber]
      properties:
        id:
          type: string
        userId:
          type: string
        studentNumber:
          type: string
        yearLevel:
          type: ["integer", "null"]
        section:
          type: ["string", "null"]
        program:
          type: ["string", "null"]
        academicYear:
          type: ["string", "null"]
        totalHoursRequired:
          type: integer
        createdAt:
          type: string
          format: date-time
    CIProfile:
      type: object
      required: [id, userId]
      properties:
        id:
          type: string
        userId:
          type: string
        employeeId:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]

    # ── STUDENT DETAILS ────────────────────────────────────────────────────
    StudentDetail:
      type: object
      required: [id, firstName, lastName, email, isActive]
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        totalHoursCompleted:
          type: number
        totalHoursRequired:
          type: integer
        attendanceRate:
          type: number
        caseCompletionRate:
          type: number
        absenceCount:
          type: integer
        lateCount:
          type: integer
        needsMakeup:
          type: boolean
        isAtRisk:
          type: boolean
    ClinicalPassport:
      type: object
      required: [studentId, totalCases, completedCases, overallCompletion, categories]
      properties:
        studentId:
          type: string
        totalCases:
          type: integer
        completedCases:
          type: integer
        overallCompletion:
          type: number
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PassportCategory"
    PassportCategory:
      type: object
      required: [category, cases, completionRate]
      properties:
        category:
          type: string
        completionRate:
          type: number
        cases:
          type: array
          items:
            $ref: "#/components/schemas/PassportCaseEntry"
    PassportCaseEntry:
      type: object
      required: [caseId, caseName, required, completed, verified, remaining, status]
      properties:
        caseId:
          type: string
        caseName:
          type: string
        category:
          type: string
        required:
          type: integer
        completed:
          type: integer
        verified:
          type: integer
        remaining:
          type: integer
        status:
          type: string
          enum: [complete, in_progress, deficient]
    HoursSummary:
      type: object
      required: [studentId, totalHoursCompleted, totalHoursRequired, hoursRemaining, progressPercent]
      properties:
        studentId:
          type: string
        totalHoursCompleted:
          type: number
        totalHoursRequired:
          type: integer
        hoursRemaining:
          type: number
        progressPercent:
          type: number

    # ── HOSPITALS ──────────────────────────────────────────────────────────
    Hospital:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        attendanceRadius:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        departments:
          type: array
          items:
            $ref: "#/components/schemas/Department"
...
    Department:
      type: object
      required: [id, hospitalId, name, isActive]
      properties:
        id:
          type: string
        hospitalId:
          type: string
        name:
          type: string
        code:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    Schedule:
      type: object
      required: [id, dutyDate, startTime, endTime, status, gracePeriodMin, createdAt]
      properties:
        id:
          type: string
        title:
          type: ["string", "null"]
        hospitalId:
          type: string
        departmentId:
          type: string
        ciId:
          type: string
        dutyDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        gracePeriodMin:
          type: integer
        status:
          type: string
          enum: [upcoming, active, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hospital:
          $ref: "#/components/schemas/Hospital"
        department:
          $ref: "#/components/schemas/Department"
        ci:
          $ref: "#/components/schemas/User"
        students:
          type: array
          items:
            $ref: "#/components/schemas/User"
...
    AttendanceRecord:
      type: object
      required: [id, scheduleId, studentId, status, createdAt]
      properties:
        id:
          type: string
        scheduleId:
          type: string
        studentId:
          type: string
        ciId:
          type: ["string", "null"]
        timeIn:
          type: ["string", "null"]
        timeOut:
          type: ["string", "null"]
        dutyHours:
          type: ["number", "null"]
        status:
          type: string
          enum: [present, late, absent, pending]
        method:
          type: string
          enum: [biometric, ci_assisted, manual]
        studentLatitude:
          type: ["number", "null"]
        studentLongitude:
          type: ["number", "null"]
        gpsVerified:
          type: boolean
        faceVerified:
          type: boolean
        livenessVerified:
          type: boolean
        remarks:
          type: ["string", "null"]
        needsMakeup:
          type: boolean
        makeupCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        student:
          $ref: "#/components/schemas/User"
        schedule:
          $ref: "#/components/schemas/Schedule"
...
    DutySlot:
      type: object
      required: [id, hospitalId, departmentId, dutyDate, startTime, endTime, maxStudents, isMakeup, status, createdAt]
      properties:
        id:
          type: string
        hospitalId:
          type: string
        departmentId:
          type: string
        ciId:
          type: ["string", "null"]
        dutyDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        maxStudents:
          type: integer
        isMakeup:
          type: boolean
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, closed, cancelled]
        createdAt:
          type: string
          format: date-time
        hospital:
          $ref: "#/components/schemas/Hospital"
        department:
          $ref: "#/components/schemas/Department"
        ci:
          $ref: "#/components/schemas/User"
        applicationsCount:
          type: integer
        approvedCount:
          type: integer
    DutySlotInput:
      type: object
      required: [hospitalId, departmentId, dutyDate, startTime, endTime, maxStudents]
      properties:
        hospitalId:
          type: string
        departmentId:
          type: string
        ciId:
          type: string
        dutyDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        maxStudents:
          type: integer
        isMakeup:
          type: boolean
        description:
          type: string
    DutyApplication:
      type: object
      required: [id, slotId, studentId, appliedAt, status]
      properties:
        id:
          type: string
        slotId:
          type: string
        studentId:
          type: string
        appliedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewedBy:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        student:
          $ref: "#/components/schemas/User"
        slot:
          $ref: "#/components/schemas/DutySlot"
    ApplicationReview:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        notes:
          type: string

    # ── RECOMMENDATIONS ────────────────────────────────────────────────────
...