# lib/api-zod/src/generated/types/dutyApplication.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ClinicalFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DutyApplicationStatus } from './dutyApplicationStatus';
import type { DutySlot } from './dutySlot';
import type { User } from './user';

export interface DutyApplication {
  id: string;
  slotId: string;
  studentId: string;
  appliedAt: Date;
  status: DutyApplicationStatus;
  /** @nullable */
  reviewedBy?: string | null;
  /** @nullable */
  reviewedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  student?: User;
  slot?: DutySlot;
}