# lib/api-zod/src/generated/types/hoursSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ClinicalFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HoursSummary {
  studentId: string;
  totalHoursCompleted: number;
  totalHoursRequired: number;
  hoursRemaining: number;
  progressPercent: number;
}